fix(query) add minimum_should_match = 1 to the "should" query (#8)
The problem appeared when the `$or` filter was used
alongside other filters (outside the `$or` group).
Expected behaviour is that at least one of the filters
in the `$or` group AND all the filters
outside the `$or` group should be met.
Elasticsearch by default returns correct results when there is only
the `$or` (Es: should) filter. However, when there are other filters
present, should seems to be ignored.

Setting `minimum_should_match = 1` fixes the problem.