refactor(parseQuery) extract special query handlers
This is to simplify the implementation of parseQuery,
making it more modular, which will help with adding more
special queries in the future.

As a bonus I - I have broken down the files into smaller,
more manageable ones.
As a bonus II - I have added eslint to dev dependencies.
Helps a bit with development.