Updates (#44)
* test(elasticsearch): add support for elasticsearch 5.6

Also, upgrade elasticsearch client to 14.1.0 and api version
to the actual elasticsearch version used in the test app.

* fix(package): remove broken start script and unknown semistandard option

* test(db): move db setup to an external script

This is to prepare for serving slightly different schema
for Es 6.0. There were some breaking changes, e.g. around
parent-child relationship. Moving schema management to a separate file
should allow the tests to stay focused and not care much
about particularites of the db config.

* test(parse-query): add missing tests and perform minor code cleanup